requirements.txt
streamlit
langchain
langchain-core
langgraph
pydantic<3,>=2
pandas
plotly
python-dotenv
httpx
alpha_vantage
fastapi
uvicorn[standard]
tavily
langchain_ollama
langchain-google-genai
fpdf2
pdfplumber
wordcloud
APScheduler
fredapi
jinja2
matplotlib
requests
